+++ src/main/java/subway/application/LineService.java
@@ -2,12 +2,12 @@ package subway.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.dao.LineDao;
 import subway.domain.*;
-import subway.dto.request.LineRequest;
-import subway.dto.request.SectionRequest;
-import subway.dto.request.StationRequest;
-import subway.dto.response.LineResponse;
+import subway.persistence.dao.LineDao;
+import subway.ui.request.LineRequest;
+import subway.ui.request.SectionRequest;
+import subway.ui.request.StationRequest;
+import subway.ui.response.LineResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
