+++ src/main/java/subway/ui/response/StationResponse.java
@@ -1,4 +1,4 @@
-package subway.dto.response;
+package subway.ui.response;
 
 import subway.domain.Station;
