+++ src/test/java/subway/integration/StationIntegrationTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.dto.response.StationResponse;
+import subway.ui.response.StationResponse;
 
 import java.util.HashMap;
 import java.util.List;
